Separate decorator from core functions
